poll_apiview 수정

+++ poll_apiview/poll/views.py
@@ -72,8 +72,8 @@ def poll_agree(request, id):
     
     try:
         if total_votes > 0:
-            poll.agreeRate = float((Decimal(poll.agree) / Decimal(total_votes)))
-            poll.disagreeRate = float((Decimal(poll.disagree) / Decimal(total_votes)))
+            poll.agreeRate = (Decimal(poll.agree) / Decimal(total_votes))
+            poll.disagreeRate = (Decimal(poll.disagree) / Decimal(total_votes))
         else:
             poll.agreeRate = Decimal(0)
             poll.disagreeRate = Decimal(0)
@@ -96,8 +96,8 @@ def poll_disagree(request, id):
     
     try:
         if total_votes > 0:
-            poll.agreeRate = float((Decimal(poll.agree) / Decimal(total_votes)))
-            poll.disagreeRate = float((Decimal(poll.disagree) / Decimal(total_votes)))
+            poll.agreeRate = (Decimal(poll.agree) / Decimal(total_votes))
+            poll.disagreeRate = (Decimal(poll.disagree) / Decimal(total_votes))
         else:
             poll.agreeRate = Decimal(0)
             poll.disagreeRate = Decimal(0)
